+++ docs/versionhistory.rst
@@ -42,6 +42,8 @@ APScheduler, see the :doc:`migration section <migration>`.
 - **BREAKING** Changed the ``timezone`` argument to ``CronTrigger.from_crontab()`` into
   a keyword-only argument
 - **BREAKING** Added the ``metadata`` field to tasks, schedules and jobs
+- **BREAKING** Added logic to store ``last_fire_time`` in datastore implementations
+  (PR by @hlobit)
 - Added the ``start_time`` and ``end_time`` arguments to ``CronTrigger.from_crontab()``
   (`#676 <https://github.com/agronholm/apscheduler/issues/676>`_)
 - Added the ``psycopg`` event broker